counsel_env/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "counsel-env"
version = "0.1.0"
dependencies = [
    "openenv-core>=0.2.1",
    "fastapi",
    "uvicorn",
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m counsel_env.server.app
server = "counsel_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["counsel_env", "counsel_env.server"]
package-dir = { "counsel_env" = ".", "counsel_env.server" = "server" }